requirements.txt
azure-ai-agentserver-agentframework==1.0.0b14
opentelemetry-semantic-conventions-ai==0.4.13
azure-ai-projects==2.0.0b3
python-dotenv
azure-identity